+++ public/js/guest/guest.js
@@ -74,24 +74,24 @@ window.onload = async () => {
       $('#lecture-status .status-text').html($('#status-starting').val());
       $('video#whiteboard').parent().addClass('running');
 
-      // joinLecture();
-      // $('#login-lecture-modal').hide();
+      joinLecture();
+      $('#login-lecture-modal').hide();
 
-      fetch(`/validate/lecture?id=${roomId}`).then((req) => {
-        switch (req.status) {
-          case 200:
-            joinLecture();
-            $('#login-lecture-modal').hide();
-            break;
-          case 404:
-            window.location.replace('/error?code=1');
-            break;
-          case 401:
-            window.location.replace('/error?code=2');
-            break;
-          default: break;
-        }
-      });
+      // fetch(`/validate/lecture?id=${roomId}`).then((req) => {
+      //   switch (req.status) {
+      //     case 200:
+      //       joinLecture();
+      //       $('#login-lecture-modal').hide();
+      //       break;
+      //     case 404:
+      //       window.location.replace('/error?code=1');
+      //       break;
+      //     case 401:
+      //       window.location.replace('/error?code=2');
+      //       break;
+      //     default: break;
+      //   }
+      // });
     }
   });
