+++ public/js/guest/guestRTC.js
@@ -135,6 +135,7 @@ export default async function initializeGuestRTC(roomId) {
                     case 'event':
                       if (typeof msg.unpublished !== 'undefined' || typeof msg.leaving !== 'undefined') {
                         // Handle here properly when the manager disconnects
+                        changeStatus.host_disconnected();
                       } else if (typeof msg.publishers !== 'undefined') {
                         joinFeed(msg.publishers);
                       }
