+++ public/js/manager/manager.js
@@ -7,24 +7,16 @@ import Whiteboard from '../classes/whiteboard.js';
 import initializeToolsMenu from '../tools.js';
 import initializeCanvasTopMenu from './canvasTopMenu.js';
 import initializeManagerChat from './managerChat.js';
-import initializeModal from './canvasModal.js';
 import initializeBoards from './managerBoards.js';
 import initializeActionsMenu from './canvasActions.js';
 import initializeManagerRTC, { changeStatus } from './managerRTC.js';
-import { getUrlId, reloadWindow } from '../utility.js';
+import { getUrlId, reloadWindow, copyTextToClipboard } from '../utility.js';
 
 const managerId = getUrlId();
 const hasAudio = $('#audioValidator').val() === 'true';
 const hasWebcam = $('#webcamValidator').val() === 'true';
 
-function beginLecture(stream) {
-  changeStatus.starting();
-  const whiteboard = new Whiteboard('canvas');
-
-  const canvasStream = whiteboard.getStream();
-
-  stream = stream || canvasStream;
-
+function beginLecture(whiteboard, canvasStream) {
   const socket = io('/', { query: `id=${managerId}` });
 
   socket.on('currentBoard', (studentSocketId) => {
@@ -53,7 +45,6 @@ function beginLecture(stream) {
     initializeCanvasTopMenu(socket, room.lecture_details.id);
     initializeToolsMenu(whiteboard);
     initializeActionsMenu(socket, whiteboard, canvasStream);
-    initializeManagerRTC(room.lecture_details.id, stream, canvasStream);
     initializeBoards(socket, whiteboard, boards, boardActive, canvasStream);
     initializeManagerChat(socket, room.lecture_details.id);
   });
@@ -62,17 +53,39 @@ function beginLecture(stream) {
 window.onload = () => {
   if (!(hasAudio || hasWebcam)) $('#modal-select-button').css('margin-bottom', '30px');
   $('#welcome-lecture-modal').show();
-  const isWebcamActive = document.getElementById('webcam') !== null;
-  const isAudioActive = document.getElementById('audio') !== null;
-  const start = (stream = null) => {
-    initializeModal(stream);
+
+  const whiteboard = new Whiteboard('canvas');
+  const canvasStream = whiteboard.getStream();
+  const roomId = $('#_id').val();
+  changeStatus.starting();
+  initializeManagerRTC(roomId, canvasStream, () => {
+    $('#modal-select-button').removeClass('live-button-inactive').find('.ld').fadeOut(function () {
+      $(this).parent().find('span').fadeIn();
+    });
+
+    $('#modal-copy-link').click(function () {
+      const copyText = document.querySelector('.modal-url-share');
+      copyTextToClipboard(copyText.innerText);
+      const range = document.createRange();
+      range.selectNodeContents(copyText);
+      const selection = window.getSelection();
+      selection.removeAllRanges();
+      selection.addRange(range);
+      this.innerHTML = 'Copied!';
+      this.style.opacity = 1;
+      setTimeout(() => {
+        this.style.opacity = 0.83;
+        this.innerHTML = 'Copy';
+        selection.removeAllRanges();
+      }, 2000);
+    });
+
     $('#modal-select-button').click(() => {
       $('#welcome-lecture-modal').hide();
-      const roomId = $('#_id').val();
       fetch(`/validate/lecture?id=${roomId}`).then((req) => {
         switch (req.status) {
           case 200:
-            beginLecture(stream);
+            beginLecture(whiteboard, canvasStream);
             $('#welcome-lecture-modal').hide();
             break;
           case 404:
@@ -85,27 +98,5 @@ window.onload = () => {
         }
       });
     });
-  };
-  if (isWebcamActive || isAudioActive) {
-    navigator.mediaDevices.getUserMedia({ audio: isAudioActive, video: isWebcamActive })
-      .then(start)
-      .catch(() => {
-        Swal.fire({
-          icon: 'error',
-          title: `<strong style="font-size: 1.2rem">${$('#swal-title').val()}</strong>`,
-          html: `<div style="font-size: .9rem; opacity: .85;">
-            ${$('#swal-text').val()}
-          </div>`,
-          confirmButtonColor: 'rgba(70, 194, 255, 1)',
-          confirmButtonText: 'Ok',
-          showClass: {
-            popup: 'animate__animated animate__fadeIn',
-          },
-          footer: `
-            <a style="color: gray; text-decoration: none;" href="https://getacclaim.zendesk.com/hc/en-us/articles/360001547832-Setting-the-default-camera-on-your-browser">
-              <i class="fa fa-question-circle" aria-hidden="true"></i> ${$('#swal-help').val()}
-            </a>`,
-        }).then(reloadWindow);
-      });
-  } else start();
+  });
 };
