+++ server/services/i18n/i18n-en.json
@@ -73,13 +73,15 @@
   "readyToJoin": "Ready to Join?",
   "goLive": "Go Live",
   "testMic": "Test Microphone",
+  "testWebcam": "Test Webcam",
   "shareURL": "Share Your Lecture URL",
   "Copy": "Copy",
   "voiceSettings": "Voice Settings",
   "inputDevice": "Input Device",
   "micTest": "Test your Microphone",
   "startMicTest": "Start a test to check up your mic!",
-  "checkMic": "Check",
+  "check": "Check",
+  "stop": "Stop",
   "pencilTool": "Pencil Tool",
   "eraserTool": "Eraser Tool",
   "paintBucketTool": "Paint Bucket Tool",
@@ -97,7 +99,8 @@
   "addNewBoard": "Add a new boards",
   "deleteCurrentBoard": "Delete current board",
   "clearBoard": "Clear Board",
-  "Mute": "Mute",
+  "Mute": "Mute Microphone",
+  "Unmute": "Unmute Microphone",
   "clickEndLecture": "Click to end lecture",
   "shareLectureLink":"Share lecture link",
   "lectureSettings": "Lecture Settings",
