+++ server/services/i18n/i18n-pt.json
@@ -71,14 +71,16 @@
   "Configure": "Configurações",
   "lectureChat": "Chat da Aula",
   "goLive": "Iniciar Aula",
-  "testMic": "Teste Microfone",
+  "testMic": "Teste o Microfone",
+  "testWebcam": "Teste a Webcam",
   "shareURL": "Compartilhe a URL da sua Aula",
   "Copy": "Copiar",
   "voiceSettings": "Ajuste de Voz",
   "inputDevice": "Dispositivos de Entrada",
   "micTest": "Teste seu Microfone",
   "startMicTest": "Comece um teste para verificar seu microfone!",
-  "checkMic": "Testar",
+  "check": "Testar",
+  "stop": "Parar",
   "pencilTool": "Lápis",
   "eraserTool": "Borracha",
   "paintBucketTool": "Preencher com cor",
@@ -96,7 +98,8 @@
   "addNewBoard": "Adicionar novo quadro",
   "deleteCurrentBoard": "Deletar quadro atual",
   "clearBoard": "Limpar quadro",
-  "Mute": "Mutar",
+  "Mute": "Mutar Microfone",
+  "Unmute": "Desmutar Microfone",
   "clickEndLecture": "Clique para terminar a aula",
   "shareLectureLink":"Copiar o link da aula",
   "lectureSettings": "Ajustes da Aula",
